Čia bandžiau padaryti, kad update'intųsi taškai:D Nežinau, ar pavyko :D

+++ API/API/UsersComponent/Models/Customer.cs
@@ -1,6 +1,6 @@
 ﻿namespace API.UsersComponent.Models;
 
-public class Customer
+public class Customer 
 {
     public Guid Id { get; set; }
     public Guid LoyaltyId { get; set; }
@@ -8,6 +8,6 @@ public class Customer
     public string LastName { get; set; } = String.Empty;
     public DateTime Birthday { get; set; }
     public string Address { get; set; } = String.Empty;
-    public int Points { get; set; }
+    public decimal Points { get; set; }
 
 }
